Added toggle to enable / disable the user email verifiaction.
The switching is done via the 'app.email_verifiaction' flag in the Firebase functions config.

No email will be send if the falg is 'false'!

This is helps to automate our testing, because we don't need to integreate a
bunch of services in our testing infrastructure and we prevent
bombarding the email services with our testing requests.

Signed-off-by: Alexander Schieweck <alexander.schieweck@lero.ie>